backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets>=13.0
google-genai==1.5.0
pillow==10.4.0
pillow-heif==0.18.0
piexif==1.1.3
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.9.0
pdfplumber>=0.11.0
